requirements.txt
six
tqdm==4.30.*
torch>=1.0
git+https://github.com/pytorch/text.git@master#wheel=torchtext
future
configargparse
sentencepiece
flask_mongoengine
indic-transliteration
kafka-python